requirements.txt
flask==2.3.3
flask-cors==4.0.0
pyyaml==6.0.1
linkedin-api==2.0.0
requests==2.31.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
selenium==4.11.2
webdriver-manager==4.0.0
